// src/app/login/models/signup.model.ts
export enum SignUpFormFields {
  NAME = 'name',
  EMAIL = 'email',
  PASSWORD = 'password',
}


export enum NameFieldErrors {
  REQUIRED = 'required',
  MIN_LENGTH = 'minlength',
}

export enum EmailFieldErrors {
  REQUIRED = 'required',
  EMAIL = 'email',
}

export enum PasswordFieldErrors {
  REQUIRED = 'required',
}